+++ src/main/java/duke/Exceptions/InvalidTaskIndexException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package duke.Exceptions;
 
 public class InvalidTaskIndexException extends LemonException {
     public InvalidTaskIndexException(String message) {
