+++ src/main/java/duke/Exceptions/LemonException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package duke.Exceptions;
 
 public class LemonException extends Exception {
     public LemonException(String message) {
