+++ src/main/java/duke/Tasks/Deadline.java
@@ -1,6 +1,6 @@
-package Tasks;
+package duke.Tasks;
 
-import Exceptions.InvalidDeadlineException;
+import duke.Exceptions.InvalidDeadlineException;
 
 import java.time.DateTimeException;
 import java.time.LocalDate;
@@ -9,7 +9,7 @@ import java.time.format.DateTimeFormatter;
 public class Deadline extends Task {
     /**
      * Taken from the Partial Solution given on https://nus-cs2103-ay2324s1.github.io/website/schedule/week2/project.html
-     * A child class of Tasks to create tasks with a deadline.
+     * A child class of duke.Tasks to create tasks with a deadline.
      */
     protected LocalDate by;
