+++ src/main/java/duke/Tasks/Task.java
@@ -1,9 +1,9 @@
-package Tasks;
+package duke.Tasks;
 
 public class Task {
     /**
      * Taken from the Partial Solution given on https://nus-cs2103-ay2324s1.github.io/website/schedule/week2/project.html
-     * A class of Tasks to create tasks that need to be in list of tasks.
+     * A class of duke.Tasks to create tasks that need to be in list of tasks.
      */
     protected String description;
     protected boolean isDone;
